+++ test/test_cubemap/main.cpp
@@ -88,7 +88,7 @@ void MainLoop()
 	static Transform cameraTrans;
 	static TransformController transCtrl;
 	static std::vector<Vertex> vertices;
-	static std::vector<u16> indices;
+	static std::vector<uint16_t> indices;
 	static MaterialPtr material;
 	static PlaneShader shader;
 	Canvas* canvas = app->GetCanvas();
@@ -134,7 +134,7 @@ void MainLoop()
 		{
 			vertices.emplace_back(Vertex{ mesh->vertices[i], mesh->texcoords[i], mesh->normals[i] });
 		}
-		for (auto idx : mesh->indices) indices.emplace_back((u16)idx);
+		for (auto idx : mesh->indices) indices.emplace_back(idx);
 
 		Rasterizer::transform = objectTrans.GetMatrix();
     }
